commits

+++ src/main/java/com/example/wmsnew/product/repository/ProductRepository.java
@@ -9,4 +9,4 @@ import java.util.List;
 
 @Repository
 public interface ProductRepository
-    extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {}
+    extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {}
